requirements.txt
sentence-transformers
faiss-cpu
numpy
PyPDF2
groq
httpx
flask
gunicorn
